Add context gathering progress tracking to agent
This commit adds comprehensive progress tracking and analysis during
context gathering searches:

- Adds SearchMetrics struct to track unique sources, query similarity,
  overlap
- Implements convergence analysis to detect repetitive patterns
- Enhances search derivation to make better stop/continue decisions
- Strengthens validation for document read operations
- Improves single-purpose focus with clear objectives
- Adds mandatory completion conditions

The main changes are:
- New SearchMetrics struct to analyze search effectiveness
- Stronger document read validation requiring IDs from listings
- Enhanced progress tracking metrics and convergence analysis
- Documentation updates for better search scope guidance

+++ queries/src/hybrid_search.rs
@@ -223,7 +223,7 @@ impl Query for FullTextSearchKnowledgeBaseQuery {
                 kbc.knowledge_base_id,
                 kbc.chunk_index,
                 kbc.content,
-                ts_rank(kbc.search_vector, plainto_tsquery('english', $1)) as text_rank,
+                ts_rank(kbc.search_vector, plainto_tsquery('english', $1))::float8 as text_rank,
                 d.title as document_title,
                 d.description as document_description
             FROM knowledge_base_document_chunks kbc
